% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{benchmarkStringDetection-package}
\alias{benchmarkStringDetection}
\alias{benchmarkStringDetection-package}
\title{benchmarkStringDetection: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Brancen Gregory \email{brancengregory@gmail.com}

}
